pyproject.toml
[project]
name = "soundscape-manager"
version = "0.1.0"
description = "Scaffold for TABby (soundscape manager)"
requires-python = ">=3.12"
dependencies = ["PyYAML>=6.0"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
